+++ wgpu-info/src/main.rs
@@ -139,6 +139,8 @@ mod inner {
         println!("\t  VendorID: {:?}", info.vendor);
         println!("\t  DeviceID: {:?}", info.device);
         println!("\t      Type: {:?}", info.device_type);
+        println!("\t    Driver: {:?}", info.driver);
+        println!("\tDriverInfo: {:?}", info.driver);
         println!("\t Compliant: {:?}", downlevel.is_webgpu_compliant());
         println!("\tFeatures:");
         for i in 0..(size_of::<wgpu::Features>() * 8) {